爬塔设定添加

+++ Assets/Scrpit/Bean/MVC/User/UserInfiniteTowersBean.cs
@@ -8,6 +8,6 @@ public class UserInfiniteTowersBean
 {
     public List<string> listMembers = new List<string>();//成员
     public bool isSend = false;//是否是派遣
-    public long layer = 0;//层数
+    public int layer = 0;//层数
     public float proForSend = 0;//派遣进度
 }
